requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==3.2.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.3
flatbuffers==24.3.25
fonttools==4.54.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
jax==0.4.35
jaxlib==0.4.35
kiwisolver==1.4.7
matplotlib==3.9.2
mediapipe==0.10.14
ml_dtypes==0.5.0
numpy==2.1.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pillow==11.0.0
protobuf==4.25.5
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
rsa==4.9
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.36
starlette==0.41.0
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0